requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.13.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
ultralytics>=8.0.0
skyfield>=1.46
pandas>=2.0.0
geopandas>=0.13.0
shapely>=2.0.0
folium>=0.14.0
exifread>=3.0.0
piexif>=1.1.3
requests>=2.31.0
tqdm>=4.65.0
pytz>=2023.3
scikit-image>=0.21.0
flask>=2.3.0
flask-cors>=4.0.0
psutil>=5.9.0